requirements.txt
aiogram>=3.4.0
asyncpg>=0.29.0
openai>=1.12.0
aiohttp>=3.9.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
